% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_format_data_to_Excel.R
\name{add_body_dft}
\alias{add_body_dft}
\title{Adds the body (main data) to the tab. Relies on many TRA25rap subfunctions. 
Not advisable to use outside of TRA25 table compilation - but useful as 
a spring board for creating bespoke functions}
\usage{
add_body_dft(tab, data_for_xl)
}
\arguments{
\item{tab}{the core tab object}

\item{data_for_xl}{output from TRA25_arrange_data}
}
\description{
Adds the body (main data) to the tab. Relies on many TRA25rap subfunctions. 
Not advisable to use outside of TRA25 table compilation - but useful as 
a spring board for creating bespoke functions
}
