% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_format_data_to_Excel.R
\name{add_hyperlink_dft}
\alias{add_hyperlink_dft}
\title{Adds a hyperlink to the second row of the sheet. Checks that is of TRA25 format
but can easily be altered not to. Based off openxlsx::writeFormula}
\usage{
add_hyperlink_dft(tab, title_text)
}
\arguments{
\item{tab}{the core tab object}

\item{title_text}{vector string with 6 elements}
}
\description{
Adds a hyperlink to the second row of the sheet. Checks that is of TRA25 format
but can easily be altered not to. Based off openxlsx::writeFormula
}
\examples{
tab <- xltabr::initialise(ws_name = "TRA2501a")
title_text <- c("Department for Transport statistics",
               "Traffic",
               "Table TRA2501a",
               "Road traffic (vehicle miles) by vehicle type in Great Britain, rolling annual totals from 1994",
               "Billion vehicle miles",
               "Rolling annual totals")
#setting the DfT style path - defines font size, colour, etc.
xltabr::set_style_path(system.file("DfT_styles.xlsx", package = "TRA25rap"))
tab <- add_title_dft(tab, title_text)
#important that following line is run BEFORE add_hyperlink_dft
tab <- xltabr::write_data_and_styles_to_wb(tab)

tab <- add_hyperlink_dft(tab, title_text)

#to see the result:
openxlsx::openXL(tab$wb)
}
