% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_format_data_to_Excel.R
\name{add_title_dft}
\alias{add_title_dft}
\title{Adds the title to the tab. title_text must be of the same format as TRA25 tables}
\usage{
add_title_dft(tab, title_text)
}
\arguments{
\item{tab}{the core tab object}

\item{title_text}{vector string with 6 elements}
}
\description{
Adds the title to the tab. title_text must be of the same format as TRA25 tables
}
\examples{
tab <- xltabr::initialise()
title_text <- c("Department for Transport statistics",
               "Traffic",
               "Table TRA2501a",
               "Road traffic (vehicle miles) by vehicle type in Great Britain, rolling annual totals from 1994",
               "Billion vehicle miles (not seasonally adjusted)",
               "Rolling annual totals")
#setting the DfT style path - defines font size, colour, etc.
xltabr::set_style_path(system.file("DfT_styles.xlsx", package = "TRA25rap"))

tab <- add_title_dft(tab, title_text)

#important xltabr function below - basically cements the formatting changes you defined
tab <- xltabr::write_data_and_styles_to_wb(tab)
#to see the result:
openxlsx::openXL(tab$wb)
}
