% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2df.R
\name{TRA25_data_api}
\alias{TRA25_data_api}
\title{Downloads data from road traffic API and formats correctly into data frame}
\usage{

  TRA25_data_api(url = "https://statistics-api.dft.gov.uk/api/roadtraffic/quarterly")
}
\arguments{
\item{url}{the API url. Preset to one that works but can be overwritten (for example, seasonal data)}
}
\description{
Downloads data from road traffic API and formats correctly into data frame
}
\examples{
raw <- TRA25_data_api()
}
