% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contents.R
\name{contents_TRA25}
\alias{contents_TRA25}
\title{contents}
\usage{
contents_TRA25(table_set, year, quarter, save_to, save_over = F,
  last_updated = "FILL IN", next_update = "FILL IN")
}
\description{
makes the contents page for TRA25 tables
}
