% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_format_data_to_Excel.R
\name{get_filename}
\alias{get_filename}
\title{Decides the name of your .xlsx file, simplest case is just use the table name.}
\usage{
get_filename(start_from_file, save_over, table_name)
}
\arguments{
\item{start_from_file}{the name of the .xlsx file that we're going to add a sheet to}

\item{save_over}{logical, if TRUE then overwrite the original file}

\item{table_name}{string, eg "TRA2501"}
}
\description{
Decides the name of your .xlsx file, simplest case is just use the table name.
}
