% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_format_data_to_Excel.R
\name{new2xl}
\alias{new2xl}
\title{new2xl}
\usage{
new2xl(data_for_xl, title_text, footer_text, table_name, save_to = getwd(),
  start_from_file = FALSE, save_over = F)
}
\arguments{
\item{data_for_xl}{data frame outputted from \code{\link{TRA25_arrange_data}}}

\item{title_text}{vector of strings, if not length 6 has warning (see Examples below for correct notation)}

\item{footer_text}{vector of strings, can be any length}

\item{table_name}{string eg "TRA2504e"}

\item{save_to}{where the xlsx document will be saved. Default is current folder.}

\item{start_from_file}{Give name of workbook the sheet will be added on to. If left blank, new workbook
will be created}

\item{save_over}{TRUE or FALSE. Should the output file replace the file of "start_from_file" or
be saved as a new file? TRUE = replace the file}
}
\description{
takes a data frame with title and footer text and outputs a beautiful Excel table.
Made with xltabr as main base.
}
\examples{
\dontrun{
#set up scenario
raw <- TRA25_data_api()
data_for_xl <- TRA25_arrange_data(raw, roll=F, type="vehicle", units="traffic", km_or_miles = "km")
#title and footer
title_text <- c("Department for Transport statistics",
               "Traffic",
               "Table TRA2504e",
              "Road traffic (vehicle kilometres) by vehicle type in Great Britain, quarterly from 1993",
              "",
              "Billion vehicle kilometres (not seasonally adjusted)")
footer_text <- footer_text <- c("",
              "[1] Two wheeled motor vehicles, buses, and coaches",
              "[2] Total may not match sum due to rounding",
              "[3] figures affected by September 2000 fuel protest",
              "The figures in these tables are National Statistics")


#apply the function (look in folder to see output)
new2xl(data_for_xl,
       title_text,
       footer_text,
       table_name = "TRA2504e",
       save_to = "/Users/Luke/Documents/xltabr_TRA2504e/sheet_builder_test",
       start_from_file = "builder.xlsx",
       save_over = F)
       }
}
