% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_original.R
\docType{data}
\name{raw_original}
\alias{raw_original}
\title{The original source raw data set as of 06/03/2018}
\format{A data frame with 1900 rows (gets updated)
and 5 variables:
\describe{
  \item{year}{}
  \item{quarter}{1 = Jan-Mar, 2 = Apr-Jun, 3 = Jul - Sep, 4 = Oct - Dec}
  \item{road_type}{}
  \item{vehicle_type}{}
  \item{estimate}{estimate in thousand vehicle kilometres}
}}
\source{
\url{https://statistics-api.dft.gov.uk/api/roadtraffic/quarterly}
}
\usage{
raw_original
}
\description{
A dataset downloaded via API
}
\keyword{datasets}
