% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_arrange_data_in_R.R
\name{TRA25_arrange_data}
\alias{TRA25_arrange_data}
\title{given the API output from \code{\link{TRA25_data_api}}, rearranges into a data frame
pivotted to the right structure for making into the Excel document.}
\usage{
TRA25_arrange_data(raw, roll = NA, type = NA, units = NA,
  km_or_miles = NA)
}
\arguments{
\item{raw}{data frame outputted from \code{\link{TRA25_data_api}()}}

\item{roll}{logical. TRUE if rolling annual values desired, FALSE if not}

\item{type}{character string. Either "road", "vehicle", or "vehicle_and_road"
dependant on which wanted for column headers}

\item{units}{character string. Either
 \itemize{
  \item "traffic" - traffic values (vehicle Kms)
  \item "percentage" - percentage change from previous year same quarter
  \item "index" - indexed values from first quarter in data, with that quarter being 100
}}

\item{km_or_miles}{"km" or "miles" dependant on desired units. Only used if units = "traffic",
as otherwise doesn't make a difference.}
}
\value{
pivotted data frame that is ready to be formatted nicely as an Excel (.xlsx) doc 
using the function TRA25rap::TRA25_format_to_xl
}
\description{
given the API output from \code{\link{TRA25_data_api}}, rearranges into a data frame
pivotted to the right structure for making into the Excel document.
}
\examples{
#first get the raw data
raw <- TRA25_data_api()
#Google TRA25 if the naming convention on the left ("TRA25...") doesn't make sense
TRA2501a_data_for_xl <- TRA25_arrange_data(raw, roll=T, type="vehicle", units="traffic", km_or_miles = "miles")
View(TRA2501a_data_for_xl) #look at the data frame created - is the same as sheet TRA2501a (search online)
}
