% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_arrange_data_in_R.R
\name{TRA25_data_api}
\alias{TRA25_data_api}
\title{Downloads data from road traffic API and makes into data frame}
\usage{

  TRA25_data_api(url = "https://statistics-api.dft.gov.uk/api/roadtraffic/quarterly")
}
\arguments{
\item{url}{the API url. Is preset to the original, but can be overwritten (for example, seasonal data)}
}
\description{
Downloads data from road traffic API and makes into data frame
}
\examples{
#set up scenario. If statement is to do with the situation when the API URL is not in place,
#with the back up (to ensure the example works) using pre-installed data
if(RCurl::url.exists("https://statistics-api.dft.gov.uk/api/roadtraffic/quarterly"))
  {raw <- TRA25_data_api()
  } else {load("~/ac-roadtraff-pipe/R/sysdata.rda")} 
}
