% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_format_data_to_Excel.R
\name{TRA25_format_to_xl}
\alias{TRA25_format_to_xl}
\title{TRA25_format_to_xl}
\usage{
TRA25_format_to_xl(data_for_xl, title_text, footer_text, table_name,
  save_to = getwd(), start_from_file = FALSE, save_over = F,
  num_dp = 1, specific_cells = F)
}
\arguments{
\item{data_for_xl}{data frame outputted from \code{\link{TRA25_arrange_data}}}

\item{title_text}{vector of strings, if not length 6 has error (see Examples below for correct notation)}

\item{footer_text}{vector of strings, can be any length}

\item{table_name}{string eg "TRA2504e"}

\item{save_to}{where the xlsx document will be saved. Default is current folder.}

\item{start_from_file}{Give name of workbook the sheet will be added on to. If left blank, new workbook
will be created}

\item{save_over}{TRUE or FALSE. Should the output file replace the file of "start_from_file" or
be saved as a new file? TRUE = replace the file. FALSE = will save with time stamp as part of name}

\item{num_dp}{number of decimal places for output. 1 is the default (and what is published)}

\item{specific_cells}{Defaults to F. This must be a list of the year, quarter, and stylename that the user
wants a specific row to be changed to. For example, specific_cells <- list(year = 2017, quarter = 4, 
style_name = "body_topline_dashed") gives that row a dashed top line. The style_name is from the styles sheet,
which in this function is system.file("DfT_styles.xlsx", package = "TRA25rap")}
}
\description{
takes a data frame with title and footer text and outputs a beautiful Excel table.
Made with xltabr as main base.
}
\examples{
\dontrun{
#set up scenario. If statement is to do with the situation when the API URL is not in place,
#with the back up (to ensure the example works) using pre-installed data
if(RCurl::url.exists("https://statistics-api.dft.gov.uk/api/roadtraffic/quarterly"))
  {raw <- TRA25_data_api()
  } else {load("~/ac-roadtraff-pipe/R/sysdata.rda")} 
 
data_for_xl <- TRA25_arrange_data(raw, roll=F, type="vehicle", units="traffic", km_or_miles = "km")
#title and footer
title_text <- c("Department for Transport statistics",
               "Traffic",
               "Table TRA2504e",
              "Road traffic (vehicle kilometres) by vehicle type in Great Britain, quarterly from 1993",
              "",
              "Billion vehicle kilometres")
footer_text <- footer_text <- c("",
              "[1] Two wheeled motor vehicles, buses, and coaches",
              "[2] Total may not match sum due to rounding",
              "[3] figures affected by September 2000 fuel protest",
              "The figures in these tables are National Statistics")

#This code below currently causes an error due to the location of start_from_file being 
#different to save_to. Is a bug that needs fixing! Has been flagged as an issue on GitHub
#as "location of files"
TRA25_format_to_xl(data_for_xl,
       title_text,
       footer_text,
       table_name = "TRA2504e",
       save_to = getwd(),
       start_from_file = system.file("template.xlsx", package="TRA25rap"),
       save_over = F)
       }
}
