% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_arrange_data_in_R.R
\name{TRA25_rolling_annual}
\alias{TRA25_rolling_annual}
\title{Ultimately a subfunction of \code{\link{TRA25_arrange_data}}}
\usage{
TRA25_rolling_annual(raw, x)
}
\arguments{
\item{raw}{the raw data, from \code{\link{TRA25_data_api}}}

\item{x}{logical, if TRUE then makes values rolling annual totals. If FALSE
then function does nothing (returns raw)}
}
\description{
Creates the rolling annual totals from quarterly values
}
\examples{
raw <- TRA25_data_api()
rolling_raw <- TRA25_rolling_annual(raw, TRUE)
}
