% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_arrange_data_in_R.R
\name{TRA25_vehicle_road}
\alias{TRA25_vehicle_road}
\title{This is a subfunction of TRA25_arrange_data}
\usage{
TRA25_vehicle_road(raw, type)
}
\arguments{
\item{raw}{the raw data in the same format as the output of \code{\link{TRA25_data_api}}}

\item{type}{either "road", "vehicle", or "vehicle_and_road"}
}
\description{
Takes the raw data, and pivots it up so that a factor variable column (with n factors)
is changed to add in the n columns, with the "estimate" col now populating the values
}
\examples{
#set up scenario. If statement is to do with the situation when the API URL is not in place,
#with the back up (to ensure the example works) using pre-installed data
if(RCurl::url.exists("https://statistics-api.dft.gov.uk/api/roadtraffic/quarterly"))
  {raw <- TRA25_data_api()
  } else {load("~/ac-roadtraff-pipe/R/sysdata.rda")} 
data_for_xl <- TRA25_vehicle_road(raw, "road")
#look at where we've gone from/to
tibble::glimpse(raw)
tibble::glimpse(data_for_xl)
}
