% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_uptodate}
\alias{check_uptodate}
\title{check_uptodate}
\usage{
check_uptodate(raw, year, quarter)
}
\arguments{
\item{raw}{output from \code{\link{TRA25_data_api}}}

\item{year}{integer - year of most recent data}

\item{quarter}{integer - quarter of most recent data}
}
\description{
checks that the incoming data's most recent info matches what the user
thinks it is
}
\examples{
mini_raw <- readRDS(system.file("data","mini_raw.rds", package = "TRA25rap"))
# can see if look at mini_raw that goes to 1995 Q4
check_uptodate(mini_raw,1995,4) #no error
\dontrun{
check_uptodate(mini_raw,1995,2) #error}
}
