% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2df.R
\name{raw2new}
\alias{raw2new}
\title{given the API output from \code{\link{TRA25_data_api}}, formats into a data frame
with values defined by the user.}
\usage{
raw2new(raw, roll = NA, type = NA, units = NA, km_or_miles = NA)
}
\arguments{
\item{raw}{data frame outputted from \code{\link{TRA25_data_api}()}}

\item{roll}{logical. TRUE if rolling annual values desired, FALSE if not}

\item{type}{character string. Either "road" or "vehicle" dependant on which wanted for column headers}

\item{units}{character string. Either
 \itemize{
  \item "traffic" - traffic values (vehicle Kms)
  \item "percentage" - percentage change from previous year same quarter
  \item "index" - indexed values from first quarter in data, with that quarter being 100
}}

\item{km_or_miles}{"km" or "miles" dependant on desired units. Only used if units = "traffic",
as otherwise doesn't make a difference.}
}
\value{
data frame of same dimensions as input, with values changed to rolling annual (if stated)
}
\description{
given the API output from \code{\link{TRA25_data_api}}, formats into a data frame
with values defined by the user.
}
\examples{
#first get the raw data
raw <- TRA25_data_api()
#Google TRA25 if the naming convention on the left ("TRA25...") doesn't make sense
TRA2504e <- raw2new(raw,roll=FALSE, type="vehicle", units="traffic", km_or_miles = "km")
TRA2505e <- raw2new(raw,roll=FALSE, type="road", units="traffic", km_or_miles = "km")
TRA2504b <- raw2new(raw,roll=TRUE, type="vehicle", units="index")
TRA2504c <- raw2new(raw,roll=TRUE, type="vehicle", units="percentage")
}
